first major re-write
removes abbreviations from the main text extraction

splits all paragraphs out individually, the logic for this has been completely re-written and looks through the HTML file in a recursive manner, looking for paragraphs, subsections or "other". Upon finding a paragraph a new paragraph is written to the output datastore, upon finding a subsection the subsection title is defined and then the recursion begins in the subsection children, if "other" is found then recursion occurs in the "other" children

Config now has a regex parameter to allow you to define how an attribute of a node should look, e.g. PMC paragraphs always have an id of P1, P2, P3 etc. The regex is used to verify that any P tags are actual paragraphs

References are not currently gathered in this commit

This commit is not thoroughly tested and is hardcoded to return the results in bioc format

It functions correctly using these arguments:

-f
Tutorial/PMC4827154.html
-t
output2
-c
configs/config_pmc.json