Update run_app.py
I added src in front of the autoCORPus when importing the package because it is in a separate folder from run_app.
I moved the assignment to out_dir to the block whose scope is to mirror the directory structure. It should be the empty string otherwise (it already was without moving it when processing a single file, but two paths are merged when processing all files in a directory, e.g. "C:/.../.../.../C:/.../..."; the merged paths were the right path, i.e. the target directory, and the path to the parent directory).
When processing all the files in a directory, the output directory and the file were separated by '\' in the path, while the rest were '/'. Therefore, I replaced '\\' with '/' everywhere and no other errors occurred anymore.
I tested my modifications on the files assigned to me.